\name{getContrasts}
\alias{getContrasts}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{q}
\description{
q
}
\usage{
getContrasts(model, sets = NULL, nsets = 1, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{model}{ q }
  \item{sets}{q  }
  \item{nsets}{ q }
  \item{\dots}{ q }
}
\details{

}
\value{
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}
}
\references{ q }
\author{q }
\note{q }


\seealso{q }
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
}
\keyword{ models }
\keyword{ regression }
\keyword{ nonlinear }
