\name{se}
\alias{se}
\title{Extract Standard Errors}
\description{
  Generic function for extracting standard errors from fitted models.
}
\usage{
se(object, ...)
}
\arguments{
  \item{object}{ A fitted model object.}
  \item{\dots}{ Arguments to methods.}
}
\value{Standard errors of model parameters.}
\author{ Heather Turner }
\keyword{ internal }
\seealso{ \code{\link{se.gnm}}}